+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## XX.XX.XX - 2023-XX-XX
 * [Changed][7999](https://github.com/stripe/stripe-android/pull/7999) In test mode, PaymentSheet now fails to load when getting saved payment methods fails.
 
+* [Added] Added support for `paymentMethodTypes` in `CustomerAdapter`, which filters payment methods to the provided list.
+
 ## 20.38.0 - 2024-02-26
 
 ### PaymentSheet